refactor: create container folder for stateful components. separate based on page
